+++ CHANGELOG.md
@@ -16,6 +16,8 @@ if they are different, the hash will be highlighted in red
 * Support for dodji (archives or comic formats) and their previews
 * Add hot keys cheat sheet in the bottom of tags view
 * Add some basic animation for hot keys sheet sheet and quick tagging
+* Add new tag pack set button for quick tagging
+* Add reset last played positions button in settings
 
 ### Viewer
 * Support opening zipped galleries: zip 7z rar cbz cbr etc It will be extracted to a temporary folder and opened